Support ovsdb-client monitor with remote connection
When using XenServer as hypervisor, we found we cannot set ovs
agent which runs in compute node with "minimize_polling=True".
Because once set it True, the ovs agent will monitor the local
ovsdb, but we want it monitor ovsdb in Dom0.
This patch is to add support for ovsdb-client to monitor remote
connection.

Change-Id: Idf2a564e96626dab3c4421a1f9139ed9ffcbcab1
Closes-bug: 1647914