Remove neutron/oslo from helpers
The enable_connection_uri() script no longer relies on cfg.CONF values or
the CLI implementation of the OVSDB API (and therefor neutron's
agent utils execute()). An executor function can be passed with kwargs so
that neutron can use rootwrap daemon, etc. If no executor is passed,
subprocess.popen/sudo will be used directly.

Neutron is the only project that uses this helper, so this change shouldn't
affect anyone else.