Remove oslo_utils dependency
Remove oslo_utils which was used for save_and_reraise exception.
In this process, I noticed that if retry=False in get_schema_helper, it
would retry anyay,  so that bug is fixed as well. The try_add_manager
attribute was added to get_schema_helper to support use cases like
OVN's where ovs-vsctl may not be installed on a machine.