Fix Python 3 compatibility in idlutils
In python 2 dict.values() would return a list while in python 3
it returns a dict_values object. The function row_by_record
tries to fetch directly the first element of the list, which
will work on python 2 but will fail on python 3 with the error
"'dict_values' object does not support indexing".

This patch addresses this issue by converting the return value
of dict.values() to list which will return the same value on
python 2 and python 3.

Change-Id: I3c39a5c6b9da3645f8833ee239820f078396fd96