Remove get_schema_helper retry/try_add_manager parameters
With enable_connection_uri taking optional arguments for handling
combinations of retrying and adding missing manager connections,
it was ugly trying to pass that information through get_schema_helper.

Neutron is the only user of the extra options to get_schema_helper and
both OVN and dragonflow have had to work around issues with
get_schema_helper and enable_connection_uri. In addition,
enable_connection_uri is Open_vSwitch schema specific. So it makes
sense to move the burden of adding any retrying or modifying
ovsdb-server connections on the user of ovsdbapp.

Change-Id: I886ccf2aac7ec5b09b47b308933b81d3d09c3284