Pass a connection to OvsdbIdl.__init__()
Originally, OvsdbIdl took a context object which in practice was
a Neutron ovs_lib.BaseOVS object. It did this because it was possible
with the old ovs-vsctl implementation to change the 'vsctl_timeout'
attribute and have that immediately reflected in the timeout value
passed to ovs-vsctl. There were some unit tests that used this. So,
when writing the CLI implementation of the OVSDB api, this context
was passed in so that the existing behavior would be honored.

With the native ovsdb implementation, this context.vsctl_timeout value
was passed to the Connection object each Transaction when it
was created. This means that the timeout, if changed on the context
object could end up being different in Connection.run and
Transaction.do_commit.

Given that non-Neutron users will never need a 'context' and that
this library does not include the ovs-vsctl implmentation, this patch
removes 'context' from the API definition and the OvsdbIdl class now
requires a 'connection' attribute and no timeout. Neutron will instead
define factory functions that fill in the appropriate attributes when
its api.get() is called.

Change-Id: Ic61c5bc988e5e2f89c92cce0a2e8169407934af9