Fixes retrieving row_by_record for ports on Windows
On Windows the name of the ports are UUIDs (for Windows
there is no 16 chars length limitation as for Linux).

Since the string is a UUID, it will not fail when it will
try to convert it with 'uuid.UUID(record)'. After this it
will try to fetch the row directly which will raise a
KeyError exception on Windows.

This does not happen on Linux because the port names
have a 16 chars length limitation. The limitation does not
apply in the Windows case. On Linux it will fail at the line
with 'uuid.UUID(record)' which will raise a ValueError
exception and will continue to lookup by other means.

This patch addresses this issue by removing the exception
RowNotFound in the Windows case.

Change-Id: Ia482a6819fbde7366b428fbcb9a05228bb637cc5