Pass the Idl directly to Connection
ovs.db.Idl does not actually create a connection on initialization.
Instead, it makes the connection when run() is called. This means
we don't have to wait until start() is called to create the Idl
instance.

In addition, adding the optional self.post_connect() call to
start() will allow networking-ovn and others to more easily use the
notification feature of the Idl. It is equivalent to the
post_initalize function in networking_ovn's OvnIdl subclass, but
as it is run after wait_for_change(), post_connect() is a more
appropriate name. A from_server() classmethod which can be used to
create an Idl instance by retrieving the schema from the server.

Change-Id: I6638bc661ae5077192c06be33e77252b91653f4c