Fix vlog level filtering
use_python_logger() receives a max_level argument which determines
the maximum level of traces that have to be logged. Current
implementation doesn't handle the max_level properly and this patch
fixes it.

Also, as the current unit test was only testing a case which worked,
this patch extends the coverage by testing also some corner cases.

Change-Id: Ic9718825fc8c97ae77bce773ecb42579f7527383