Add DNS APIs
OVN Northbound DB has DNS table. Support CRUD operations for
this table. These APIs will be used by networking-ovn to use
OVN's native DNS feature.

Note: ovn-nbctl do not support these APIs as of now. But it is
good to have these in ovsdbapp and it will be useful.

Partial-bug: #1688172

Change-Id: I5528829eef0a7534d9266bb3abc2aec269e4ba35