Add set_column(s) for ovs_idl backend
The create/set Commands allow one to pass the result of txn.add()
to a subsequent Command in a transaction to use the result of that
Command's execution. In this way one can create a DHCP Option and
then pass it to an lsp_add() command.

This change replaces the direct uses of setattr() with a version
that does the substituation if one of the arguments is a Command
instead of a normal value, enabling this behavior for all Commands.

Change-Id: I97e7cbff7e71f7e1f00342c724af963a411b6c5d