Ensure idl.run() called on TRY_AGAIN
If the DB connection drops after Connection.run() calls idl.run()
but before Transaction.do_commit(), then it is possible for
idl.commit() to return TRY_AGAIN due to the failure of session.send().
This prevents commit_block() from running idl.run() which would
actually trigger the DB reconnect, causing each successive call to
commit_block() to fail in the same way until we stop trying again
due to timeout and Connection.run() calls idl.run() again.

For now, we can just ensure that idl.run() is called when we get a
TRY_AGAIN. It might be worth it to refactor this a bit to not even
use commit_block().

Change-Id: I77402bde3fc4d45c770b9f48882870cfc91d719d
Closes-Bug: #1741889