Transaction: Handle NOT_LOCKED status
This patch is changing the commit_block() from the Transaction class to
handle the NOT_LOCKED status and treat it as an error.

The reason to treat it as an error is because, when the connection with
the OVSDB is reestabilished we don't know if that transaction will
belong to a worker that is holding a valid lock or not, so we won't
attempt to retry it.

The patch is also adding a log for unknown status, in case another
status that is not handled by ovsdbapp is returned the log will give us
some hints about what's going on instead of failing siently.

Related-Bug: #1754291
Change-Id: I5dcb226c387092cba1adb9dbf1ec782e0bac66ba
(cherry picked from commit b7c9dc9f0a86ddce99c8b189614bc4e39e0c59ff)