Ensure timeout on queueing transaction
If idl.run() somehow throws an Exception, the existing code
would cause the Connection thread to die and then if more than
one transaction was queued, it would hang forever. Instead, we
enforce timeout on queing a transaction and log the exception
that happens in the Connection thread to make debugging easier
and try reconnecting on the off chance that the exception is
recoverable.

Change-Id: I14313df9a59424787230b3fbd630d925323969fe
(cherry picked from commit c3d639ab45f827a19429d107f373fdbd555b4bfb)