Improve DbListCommand operation from O(n^2) to O(n)
Right now the DbListCommand retrieves the uuid's of all the elements
passed in as argument. This is an O(n^2) operation so when the number
of elements in a grows it's likely that we get Timeout Exceptions.

Instead of doing this, whenever possible, we'll retrieve all the
elements (from the in-memory replica) and only fetch those who were
passed as arguments avoiding the O(n^2) operation.

(cherry picked from 476b174f7d303c501fa2f3b91efda659ed9b9f3d)
Change-Id: I71411e5dd68753eb3825f8b0e91009f87de1b260
Closes-Bug: #1769897