Improve DbListCommand operation from O(n^2) to O(n)
Right now the DbListCommand retrieves the uuid's of all the elements
passed in as argument. This is an O(n^2) operation so when the number
of elements in a grows it's likely that we get Timeout Exceptions.

Instead of doing this, whenever possible, we'll retrieve all the
elements (from the in-memory replica) and only fetch those who were
passed as arguments avoiding the O(n^2) operation.

NOTE: this cherry pick conflicted because in stable/pike, the commands.py
file was in different location.

Closes-Bug: #1769897
(cherry picked from fd64d41ea6f208a5752fbcd45952a04633ee1392)
Signed-off-by: Daniel Alvarez <dalvarez@redhat.com>

Change-Id: Iacded885bb852315263ee926b9af5a1c3dc0dbba