fix tox python3 overrides
We want to default to running all tox environments under python 3, so
set the basepython value in each environment.

We do not want to specify a minor version number, because we do not
want to have to update the file every time we upgrade python.

We do not want to set the override once in testenv, because that
breaks the more specific versions used in default environments like
py35 and py36.

Change-Id: I2a995f93b7e2a4a3ad40856773ce3ba7c6666720
Signed-off-by: Doug Hellmann <doug@doughellmann.com>
Signed-off-by: Chuck Short <chucks@redhat.com>
Depends-on: Ia7412f8ec97cae2e0c8eda84eb38822a110c1a65