NBDB API: Add param if_exists for methods using db_remove()
This patch is adding the parameter "if_exists" to the methods in the
Northbound DB API that uses the db_remove() method underneath.

The following methods were affected:

* dns_remove_record()
* ls_remove_dns_record()

The intention of this patch is to make the API more flexible for
projects relying on those methods so they can choose whether they want
to raise an exception if the item being removed exist or not.

Closes-Bug: #1793132
Change-Id: I1e6749a9f3827f91a9db29e6f5ba0c8b7a6cc622