Add WaitEvent to the API
The ability to wait on an event to complete is a useful construct.
Instead of just having a one-off class to do it in the functional
tests, add it to the formal API.

There were some pylint issues with unused variables that are also
fixed in this patch.

Change-Id: Ie6d7c2ab384e152776b3d5f000c4ee405a8a3974
(cherry picked from commit 0e5904ac9e1f334ce0a7363d7778e3a3cadd67e0)