Expand retry behavior to cover other python-ovs methods
If there is a disconnect, it is possible that we could throw an
exception if we call wait() while the stream is currently set to
None. Move the try: earlier so that if something weird happens, we
at least retry before raising.

Change-Id: Ic820ab3978a3b58a41762e6fca034f62d51a0243
(cherry picked from commit 46e4323bed91b6b5eb2779ec62754d9486237590)