Add Gateway_Chassis support
This adds basic support for adding Gateway_Chassis entries as part
of adding a Logical_Router_Port. Since Gateway_Chassis is not a
root table, creating one that wasn't linked to a LRP would be a
NoOp, so there are no API methods for creating/deleting GCs. They
will automatically be deleted by ovdsb-server when the parent LRP
is deleted.

Change-Id: I12d033fac8bcbed73f8de85c688d7cd18c8bdbfb