Attempt to fetch the schema from any remote
The current implementation of the get_schema_helper function assumes one
remote is provided. When ovsdbapp is used to connect to a clustered
ovsdb, multiple comma separated remotes may be provided in the
connection argument of get_schema_helper. This function must therefore
attempt to fetch the schema from any of the provided remotes. This
change implements this functionality.

Change-Id: Ib4614401632e64421952777a60920796179d602f