Allow read-only Commands to bypass txns in execute()
Since the ovs_idl backend maintains an in-memory copy of the DB,
read-only commands can be executed w/o creating a transaction.
Since it is possible for a function that calls execute() to be
automatically nested into an existing transaction, execute()
wasn't able to access the result since it only happens once the
parent transaction completed. Handling these read-only commands
w/o creating a transaction allows the results to be read and
avoids wasting resources by creating needless transactions.

ReadOnlyCommands can still be used in with transaction(): blocks
with no changes.

Change-Id: I62d6fc2a209535227686a2397b8f7cb169e7d35b
Closes-bug: 1813213