Add db_create_row method
Since there are row-returning versions of db_find/list, and
db_create currently returns a uuid, it makes sense to have a
version that returns a RowView as well. I've seen some usages
where people do a db_create() followed by a db_find() to get the
row, and one shouldn't have to do that.

Change-Id: I5250e37a54411297a64568a3e8135855f06de807