Make it possible to opt out of nested transactions
Nested transactions work for simple things, but intermediate
results from transaction commands like create/add that are used
in subsequent nested transactions won't exist until the the end of
the whole nested transaction, causing failures. Since this can be
surprising, make it possible to easily opt out on a per-transaction
or per-API basis.

Change-Id: I4f7997d5dc11a190e188c16e8eac9fe5df8b0d0d