Allow match_ip in LrNatDelCommand to be IP network
__init__ of LrNatDelCommand in schema/ovn_northbound/commands.py takes
optional match_ip parameter to find one or more specific NAT Table rows
of a logical router to delete. This match_ip is searched in the
"logical_ip" column for NAT rows of type "snat" and in "external_ip"
column for other types. While "external_ip" column value has to be an
IP address, "logical_ip" column value can be IPv4 network or
IP address. Currently the code only allows match_ip to be an IP address
and raises an error when an IP network is passed.

This change allows match_ip to be IP network when NAT type is "snat".

Closes-Bug: #1825740
Change-Id: I90dfdb5b5fe625ca8af768dfe9dac10f6d3c1430