Add HA Chassis Group related commands
This patch is introducing new commands to ovsdbapp to work with the new
HA_Chassis_Group and HA_Chassis tables in OVN Northbound database.

These tables were introduced by the commit [0] in core OVN which created
a new type of port called "external" that can be used for various things
including but not limited to: SR-IOV and baremetal booting support.

[0]
https://github.com/ovn-org/ovn/commit/b31c76000bef314b68e776d318d1ce4cf152450b

Change-Id: I77ad33223607bee63a96e548f308bbffc1c9de52
Signed-off-by: Lucas Alvares Gomes <lucasagomes@gmail.com>