Make it possible to reference a newly created object within a transaction
This is a continuation of https://review.opendev.org/#/c/358101/

Enable the OVSDB API implementations to handle referencing a newly
created object within the same transaction. This change extends the
functionality for records that are actually commands from a previously
added transaction. Example:

   with api.transaction(check_error=True) as txn:
       sw = txn.add(api.db_create_row('Logical_Switch', name="sw0"))
       prt = txn.add(api.db_create_row('Logical_Switch_Port', name="sw0-port1"))
       txn.add(api.db_add('Logical_Switch', sw, "ports", prt))
       txn.add(api.db_add('Logical_Switch_Port', prt, "addresses", "50:54:00:00:00:01 10.0.0.10"))

Also added functional test to exercise this codepath.

Change-Id: I434707bd798d6640f0ab25bc6e29b941230862b9
Closes-Bug: #1615105
Signed-off-by: Flavio Fernandes <flaviof@redhat.com>
Co-authored-by: Terry Wilson <twilson@redhat.com>
Depends-On: https://review.opendev.org/358101/