Call row.verify() when updating map columns with db_set
The verify() method will ensure that if the row's column in the db
has changed since we set it, that we will be notified and try the
transaction again with new values. This is important when we do a
read-update operation like we do for map columns in db_set. Since
we read the values add a few keys, then overwrite the whole value,
it is important that we don't do that if we have stale data.

Since ovs 2.6, there has been a setkey() operation that would
allow modifying a map w/o overwriting the whole value, but there
is a bug where it doesn't handle maps that have a refTable value
like Qos.queues. So using that before it is fixed [1] and all
packages have picked it up would cause breakage.

[1] https://patchwork.ozlabs.org/patch/1254735/

Closes-bug: #1867652
Change-Id: Ic898cefb4ced76e9e169b9f5d6caaace600abb0c
(cherry picked from commit b92d8742bf3f5e436bcb9160618b343e3c4595fa)