Don't let an Exception in one event affect others
A RowEvent that is added that raises an exception from matches()
would cause the loop to abort that checks whether that event
matches successive RowEvents. Since it is fairly common in UPDATE
events to want to check old.column != row.column etc. to test
whether a RowEvent should fire and 'old' only contains columns that
have been modified, it seems reasonable to treat Exceptions as
"not matching" and to debug log enough info in case there was an
unintended error.

Change-Id: I0cda6044ee7652350369ff4d5d2b0ab8289c127e