OVNNB backend: Fix LbDelCommand when using vip param
Fix issue in ovn_northbound LbDelCommand where logic
for removing vip while using if_exists=True was not
doing the right thing.

In order to be consistent with as ovn-nbctl [1], this
change also introduces a modified behavior. When the
last vip is removed, it will also remove the load
balancer instance.

[1]: https://github.com/ovn-org/ovn/blob/1e2f4aaab67090935bba6347358ad9c56b0868bf/utilities/ovn-nbctl.c#L2870

Closes-Bug: #1877673
Change-Id: Ie8ae8f5efae2afb958ed2010053ff04372cc1c51