Add QoSDelExtIdCommand
This command removes all QoS rules matching a set of keys and
expected values passed in the command instantiation. No exception
is thrown if no QoS is found with the required external IDs.

For example, this will be used in Neutron to delete all QoS
registers related to a floating IP, searching for
neutron.common.ovn.constants.OVN_FIP_EXT_ID_KEY key and
the floating IP register UUID.

Change-Id: I0eb7151eabaf40232f79797b2775c9bb7891013f
Related-Bug: #1877408