Handle index lookups in DbFindCommand
This adds index lookup support to DbFindCommand. If multiple
conditions are passed that match different indexes, it will use
the intersection of the results to limit the search space.

Still needed would be support for multi-column indexes which we
currently don't use anywhere. At some point it starts looking like
a query optimizer.
Change-Id: I22965ef38c1e0bfdfa79389a265bac12cafcc28b