Add MAC_Binding index lookup
This is required because MAC_Binding does not have a "simple" one-column
index in the OVSDB schema, so it will not have an autocreated index.
Neutron code does a db_find of the MAC_Binding entry by the "ip" column,
so adding this entry allows us to do an indexed lookup.

Change-Id: I0250c7bd51c63fea534c3b6b55bb5bb053637b19