Add lock when calling idl.run()
python-ovs transparently handles DB reconnections during calls to
Idl.run(). Since ovsdbapp uses a separate thread for the
Connection (so it can handle responding to OVSDB keep-alive echo
requests) and for processing RowEvent notifcations, it is possible
that a thread could be trying to access OVSDB data at the same
time that the Connection thread is updating it. Especially during
a reconnect where the in-memory copy of the DB will be completely
re-written, this causes a problem.

Theoretically, an application that uses ovsdbapp shouldn't access
anything in the backend code directly, but since there is now only
one supported Backend, that restriction has slipped a bit. If
all accesses were in Command objects, mostly there wouldn't be a
problem as they would be handled in the Connection thread. But
with both ReadOnlyCommands and use of lookup() and anything in
idlutils, the main thread can access the in-memory db directly and
potentially cause problems.

The most common problematic method to be called outside of Command
objects is lookup() and ReadOnlyCommands. This patch adds a lock
around calls to idl.run() and lookup() and execute() for the
ReadOnlyCommand case. row_by_value() is another target, but it is
left untouched because it doesn't have access to the API instance
and it is *mostly* called from inside Command objects. User code
that doesn't will be easy to find because user code basically
shouldn't use idlutils.

Conflicts:
  ovsdbapp/tests/unit/backend/test_ovs_idl.py

Change-Id: I98c37771883103e1fb0468de9cf85364071993fa
Closes-Bug: #1888878
(cherry picked from commit 3cf8a427c0a4dfc248e0778c3b1ee65671d1d07b)