Don't give up when an Exception happens in idl.run
It's possible that idl.run() could have a bug where it raises an
Exception for an extended period of time while ovsdb-server is
down, but recover once ovsdb-server comes back up. Specifically,
python-ovs currently doesn't properly catch an exception when the
socket type is 'ssl' that it catches for other protocols.

Change-Id: Ia068650d2db3d5d8642771a6df5a260d692aea20
Closes-Bug: #1895727