Avoid race condition with RowEvent handling
Row objects are inherently tied to the transaction processing of the Idl.
This means that if you have a reference to a Row in one thread, and
another thread starts a transaction that modifies that row, the Row can
change w/o you knowing it. This is especially noticeable when using the
RowEventHandler. It is possible for a Row that is passed to notify() by
the Idl class to change between being matched and the RowEvent.run()
method being called.

This patch returns an immutable representation of the row by using the
same class that custom indexes use for searching. This should be safe
to pass to other threads.

Conflicts:
    ovsdbapp/backend/ovs_idl/idlutils.py

Change-Id: Iff6e9fdfe032e1c007e35fc7b018114e66acc895
Closes-Bug: #1896816
(cherry picked from commit dc09d6696fe175e53c0c1e4a4263d1a3e1513c9c)