Add update_tables method to Idl
The method adds given tables to the current connection schema. If the
table is not already there, it's added. If the table in current
connection doesn't exist in the provided schema, it's not added, which
seems like it's been removed.

This behavior is handy for cases where ovsdb-server is changed to new
schema.

Related-bug: #1901527

Change-Id: Ib071dd0f4d9c6e08fafc282c7551086de4b59500
Co-Authored-By: Terry Wilson <twilson@redhat.com>
Signed-off-by: Jakub Libosvar <libosvar@redhat.com>