Add an active wait in the "Backend.lookup"
Added an active wait in the method "Backend.lookup". If a timeout
and a RowEventHandler instance, listening on the same IDL connection,
are passed, if the "lookup" method does not find the requested
register, the method creates a WaitEvent instance related to the
table and register requested.

If the wait event returns, the expected record will be stored in the
event "result" member and will be the value returned by this method.

If the wait event timeout expires, the method will throw an
``idlutils.RowNotFound`` exception.

Closes-Bug: #1922934

Change-Id: I580bfb640d8157b4e749d7b1fe5de7568e99943d