Don't spam retries 100s of times a second
We can get a TRY_AGAIN because the DB is down or because the
a command is doing an update that will never pass the verify()
condition. If this happens, we shouldn't send 100s of requests per
second until we eventually time out or things come back up, so add
a simple exponential backoff with reasonable max retry sleep of 8s.

Closes-Bug: #1907836
Change-Id: Ib757f182a4c7fdf25ed9374963ac5210e4d35523
(cherry picked form commit 6596164f51217cc7fabf302ce14ccc9d9beaff1f)