Actually close the connection in Connection.stop()
Although we were stopping the connection thread, we never actually
close the Idl's connection to ovsdb-server. This is especially
noticeable in neutron's "short-lived" connections for ensuring the
creation of the pg_drop Port_Group which happen pre-fork and so
the sockets are getting cloned after the fork since they are still
open.

Closes-Bug: #1929119
Change-Id: I8a57c73f20a299b811ee31b55d40a76de9deb1cd