Check WaitEvents for match after other events
It is useful to be able to have an Event that takes an action
when matched on a particular condition, e.g. "create a neutron
agent when a Chassis is created" and to be able to create a
WaitEvent that would fire on the same condition, *after* the other
event is finished. E.g. "after the neutron agent is created, call
get_agent() to verify that it is there."

Since RowEventHandler._watched_events is a set, events are not
ordered, so sometimes the WaitEvent matches before the other
event, and things fail.

This patch automatically adds a priority queue, with RowEvent and
WaitEvent assigned appropriate priorities that will ensure that
WaitEvents are processed later than RowEvents.

Co-authored-by: Jakub Libosvar <libosvar@redhat.com>
Change-Id: I14980aa57541962742faf4ca72ed0bd3ce6e20e3