Remove ovsdb_connection singleton for tests
Since ovsdb_connection is a class attribute, different tests can
step on each other's connection. This can lead to the venv setting
up processes to connect to one server, while the test is actually
writing to another. In the lsp_bind tests where we wait on a port
binding to be created, this can be especially problematic.

Change-Id: Ic72fce11a5c16c25946f55d07d7af9e34ef283c3