Revert "Add an active wait in the "Backend.lookup""
This reverts commit bc004045cd30a708eadd64ded1af1b30ef1d1918.

There turned out to be some unresolvable issues with waiting for
events when lookup() is called from within a Command object in
the connection thread. Also the lookup of the notify_handler
was looking on the wrong object, so unless the notify_handler was
specifically passed (and we are not aware of any code that did),
this lookup timeout never actually was used.

Removing the feature, but leaving the arguments to avoid breaking
any code and adding a warning to update code if they are passed.

Change-Id: I42072ecb6a658814c7ff48cbc6f2a37a91d1eba8