Use OVN's OVS submodule for functional tests
The OVN tree for several years has had an OVS submodule that pulls
in the version of OVS it is guaranteed to work with. There is
currently a patch in OVS master that breaks the build of OVN main
so switch to using the submodule.

Change-Id: I61195cd3ebf8a5015890d3ba3988c3d9795ee817
(cherry picked from commit ee28e381fb46a04be3451de65a5166dd457eede8)