Add if_exists and **kwargs columns to db_set
The original API for db_set() out of an over-abundance of caution
use *args tuples in the format of ('column_name', value) due to the
fear that there may be OVSDB column names that were not valid
Python argument names (e.g. 'some-column'). But the OVSDB RFC
forbids column names that would be a problem.

This patch adds **columns which will be used only if there are
no values passed as *col_values. It also adds if_exists support
to db_set(), as there are situations where you don't want to throw
an error if you are updating a Row in a multi-op transaction and
it's possible that the row has been deleted on the server.

Change-Id: Ia17e545706a16f5c726ec52bae339e9362b0b614