Remove ovsdbapp-tempest-dsvm-networking-ovn-ovs-release job definition
Parent of this job (networking-ovn-tempest-dsvm-ovs-release) was defined
in the networking-ovn repo in stable/train branch which is now EOL.
Because of that this is now causing zuul config error as
networking-ovn-tempest-dsvm-ovs-release don't exists anymore.
Lets delete ovsdbapp-tempest-dsvm-networking-ovn-ovs-release job
definition and keep just basic jobs in stable/train branch.

Additionally this patch pinned OVS and OVN branch in the functional
tests job to the same versions as were used in the networking-ovn jobs
in stable/train branch.

Change-Id: If5bb7e03cb5098f34c29ee620867c341d3ec078f