Handle events with conditions and match_fn
The original goal of adding match_fn() the ovs_idl backend
implementation of RowEvent was to give an option to specify more
complex matching behavior than conditions/old_conditions allowed.

It was never intended that users would use both together, but it
also was not documented that that was the case. This patch allows
conditions and match_fn() to be treated as a logical AND to avoid
accidentally removing conditions when adding a match_fn()

Change-Id: Ifc18bf42e2a0ce8247bf986d14646f66d9554ca7