Fix BFD functional test, avoid comparing status column
The status column is updated by OVN so we cannot rely on its
contents being stable.

The test already attempts to address this by using frozen rows,
but that does not fully solve the issue as objects are created at
different points in time and then compared, which may give
unpredictable results.

Filter out the `status` column from the affected tests.

Closes-Bug: #2058264
Change-Id: Ibaa54d93109434b445f54de583f534021defccbb
Signed-off-by: Frode Nordahl <fnordahl@ubuntu.com>