[CI] Fix multiple gate issue
1. Use OVS main branch

git checkout of OVS master branch causes an error [1].
OVS deleted its master branch and replaced it with main branch.

2. Cap setuptools <71.0.0

py38 jobs (on ubuntu-focal) started to fail (see the relevant github
issue [1]) due to recent virtualenv release (20.26.4; which bundles
setuptools).

setuptools is bundled in virtualenv, so it has to be capped via the
virtualenv package. tox also needed to be capped (<4) as gate uses
tox 3.28.0, but with capping virtualenv we pull in latest tox as well,
which would cause other errors.

3. Remove lower-constraints job

lower-constraints jobs are not supported anymore and are removed from
most of the branches, so let's remove here as well to unblock the gate.

[1] error: pathspec 'master' did not match any file(s) known to git
[2] https://github.com/pypa/setuptools/issues/4483

Change-Id: I010e145f4c5079a80b72c5229dfacc4d6896373e