Add Connection.force_reconnect
The python-ovs Idl class has a force_reconnect() method, which if
called when using ovsdbapp will likely not cause a reconnection
until some other transaction or database happens due to the
Connection thread being in a poller.block() call.

This adds a Connection.force_reconnect() method that will call
Idl.reconnect and alert the poll to wake up.

Change-Id: Ie84ed53761a206f5c230e15f0e4af5710560c116