Improve Row stringification
Add the Row's uuid and support for handling converting Rows that
are part of sets and maps to UUIDs instead of logging the repr
which is the default Python pointer info.

python-ovs has a Row.__str__ and this is changing there as well.
Once the version that has it is our minimum required version,
row2str() can just return str(row).

Change-Id: I96ead59a0b278ab87a2548f7bff1252e1071dfbb
Signed-off-by: Terry Wilson <twilson@redhat.com>