Re-apply Improve Row stringification
Add the Row's uuid and support for handling converting Rows that
are part of sets and maps to UUIDs instead of logging the repr
which is the default Python pointer info.

python-ovs has a Row.__str__ and this is changing there as well.
Once the version that has it is our minimum required version,
row2str() can just return str(row).

This reapplies the patch that was reverted as part of
61cde4703cf032c7ce6bd8dce10c5822dbe12b09, after fixing the
its parent patch.

Change-Id: I13fcdaf009e59d16c44e2af73c5cef89bf88f5e5
Signed-off-by: Terry Wilson <twilson@redhat.com>
(cherry picked from commit 7744650d02062b99115535fb02108b6eeab6f7b9)