Introduce event_name property
The event_name is typically set manually as a class attribute and
the value is the same as the class name. That requires to re-set the
class attribute whenever a new subclass is created.

The patch makes the event_name a property that always calls to its own
class name so we do not need to explicitly name the event.

If a different name than the class name is needed, we can always
override the property itself.

Change-Id: I1a62a2e1b4f4aeecf0ec9698f7b55ed0bed73a5a
Signed-off-by: Jakub Libosvar <libosvar@redhat.com>