Improve ``get_uuid``, now accepts RowView|Row
The methdo ``get_uuid`` now accepts ``RowView`` and ``Row`` input
values. The UUID field is a member of these classes.

In general, the ``get_uuid`` method retrieves the object ``uuid`` field,
if present, or returns the tested object itself. In case of
``BaseCommand`` derived objects, the method first retrieves the
``result`` value, that should store a ``RowView|Row`` class object.

Signed-off-by: Rodolfo Alonso Hernandez <ralonsoh@redhat.com>
Change-Id: I903334f4566ac6dd28452a21cde5ae96fd9a26e3